Transfer code from bachelor thesis repository.
This code already existed in my bachelore thesis repository, but it was tinkered to only work for ReversiXT. The code was copied and all references to ReversiXT where replaced by generic components.
The code is not fully testet and has some weakneses, but it is a solid base to start training an ai for a generic turn based game.